extract sequencer logic from cherry pick lib, since it also applies to revert
